move components to different folder, added stories
